refactor(pdf-service): move pdfFromUserData to shared application logic
